Fixed Random Game bug, Deleted excess files.
Deleted the .csv and .txt files which held the data for the DB. I am
going to enter this in manually in my database application to ensure
that the data will not be tampered with again. I'm adding a save feature
tomorrow to create backups of the DB. Also fixed a small bug with the
random game feature, now it correctly chooses the random games.