Added Save/Load/Delete Features
Added the ability to save you text area in a file, load that text area
into the game list, and also delete the data if something goes wrong. I
have to fix one or two more bugs with this to make it easier to use.
Also, need to update text area accordingly so users don't have to exit
app.